pyproject.toml
[tool.poetry]
name = "ankipetproject"
version = "0.1.0"
description = ""
authors = ["rostykpython <89851597+rostykpython@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
genanki = "^0.13.0"
tk = "^0.1.0"
customtkinter = "^5.1.2"
openai = "^0.27.2"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
